chore: move remove_exchange_credentials to config module
