+++ freqtrade/configuration/__init__.py
@@ -1,6 +1,6 @@
 # flake8: noqa: F401
 
-from freqtrade.configuration.config_secrets import sanitize_config
+from freqtrade.configuration.config_secrets import remove_exchange_credentials, sanitize_config
 from freqtrade.configuration.config_setup import setup_utils_configuration
 from freqtrade.configuration.config_validation import validate_config_consistency
 from freqtrade.configuration.configuration import Configuration
